refactor(via-router): additional code cleanup
